#48: Fixed Settings Crash Bug
Fixed a bug that caused the app to crash when attempting to load the Settings Activity. Apparently, every activity needs to be present in AndroidManifest.xml if you want to use it. This also means that we have to apply the theme for every Activity manually unless we want them all to follow the theme of the splash screen.